+++ codeFonctionnelle10-06-2021/codeFonctionnelle.cpp/consigne.h
@@ -8,7 +8,7 @@ class consigne
   int InitAnalog();
   
   int put_hors_gel(bool hors_gel);
-  int consigne_de_chauffe(bool hors_gel , int tempActuel, float consigneRad, int heure, int power);
+  int consigne_de_chauffe(bool hors_gel , float tempActuel, float consigneRad, int heure, int power);
 
   private:
